requirements.txt
fastapi
uvicorn[standard]
pydantic
PyJWT
python-multipart  # IF UPLOADING FILES
python-dotenv
